Implement protocol specific scheduler operations
