+++ plt/plt-scheduler/src/locks/lock_controller_simple.rs
@@ -1,5 +1,7 @@
-use concordium_base::protocol_level_tokens::CborHolderAccount;
+use concordium_base::base::AccountIndex;
+use concordium_base::protocol_level_tokens::{CborHolderAccount, TokenId};
 use plt_block_state::block_state_interface::BlockStateQuery;
+use plt_block_state::entity::block_state::TokenNotFoundByIdError;
 use plt_block_state::external::AccountNotFoundByIndexError;
 use plt_block_state::persistent::protocol_level_locks::p11::{
     LockControllerSimpleV0, LockControllerSimpleV0Grant,
@@ -7,7 +9,6 @@ use plt_block_state::persistent::protocol_level_locks::p11::{
 use plt_scheduler_types::types::reject_reasons::TransactionRejectReason;
 
 use crate::locks::lock_controller::{LockController, LockControllerRejectReason, LockOperation};
-use crate::scheduler::helpers::{lookup_account_index, lookup_token_id};
 
 impl LockController for LockControllerSimpleV0 {
     // TODO: implemented as part of COR-2305
@@ -82,5 +83,39 @@ impl LockController for LockControllerSimpleV0 {
             keep_alive: config.keep_alive,
             memo: config.memo,
         })
+    }
+}
+
+
+/// Look up the account index for a [`CborHolderAccount`]. If the account does
+/// not exist, a [`TransactionRejectReason::InvalidAccountReference`] is
+/// returned.
+pub fn lookup_account_index<BSQ: BlockStateQuery>(
+    block_state: &BSQ,
+    holder: CborHolderAccount,
+) -> Result<AccountIndex, TransactionRejectReason> {
+    match block_state.account_by_address(&holder.address) {
+        Ok(account) => Ok(block_state.account_index(&account)),
+        Err(_) => Err(TransactionRejectReason::InvalidAccountReference(
+            holder.address,
+        )),
+    }
+}
+
+
+/// Look up the token ID in the block state. If the token ID does not exist, a
+/// [`TransactionRejectReason::NonExistentTokenId`] is returned.
+/// Otherwise, this returns the canonical representation of the token ID.
+///
+/// TODO: Consider returning token index instead.
+pub fn lookup_token_id<BSQ: BlockStateQuery>(
+    block_state: &BSQ,
+    token_id: TokenId,
+) -> Result<TokenId, TransactionRejectReason> {
+    match block_state.token_by_id(&token_id) {
+        Ok(token) => Ok(block_state.token_configuration(&token).token_id),
+        Err(TokenNotFoundByIdError(_)) => {
+            Err(TransactionRejectReason::NonExistentTokenId(token_id))
+        }
     }
 }
