+++ plt/plt-scheduler/src/scheduler.rs
@@ -1,14 +1,8 @@
 //! Entry points to calling the scheduler. The scheduler is responsible for executing
 //! transaction and update instruction payloads.
 
-use crate::transaction_execution::TransactionExecution;
-use concordium_base::base::{Energy, Nonce, ProtocolVersion};
-use concordium_base::contracts_common::AccountAddress;
-use concordium_base::transactions::Payload;
-use concordium_base::updates::UpdatePayload;
 use plt_block_state::block_state_interface::BlockStateOperations;
 use plt_block_state::failure::BlockStateFailure;
-use plt_scheduler_types::types::execution::{ChainUpdateOutcome, TransactionExecutionSummary};
 
 pub mod helpers;
 mod p11;
